Add "abstractClass" support to prototype chain
This enables extensions to each have a base model and
view for providing common functionality.  The abstract
classes shouldn't need as much decoration as a full-fledged
class though.

Also remove the function.name polyfill for IE.  Since the
names are all changed when the JS is minimized, there's not
much point on relying on it.